Refactor: Use decimal for division to improve precision

+++ src/Calculator/BasicArithmetic.cs
@@ -4,11 +4,11 @@ public static class BasicArithmetic
 {
     public static int Sum(int left, int right) => left + right;
     public static int Multiply(int left, int right) => left * right;
-    public static double Divide(int left, int right)
+    public static decimal Divide(decimal left, decimal right)
     {
         if (right == 0)
             throw new DivideByZeroException("Cannot divide by zero.");
         
-        return (double)left / right;
+        return left / right;
     }
 }
